requirements.txt
Flask
openai
twilio
textblob
python-dotenv
